# web/src/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "ScrapLink",
    "version": "0.1.0"
  },
  "paths": {
...
    "/admin/disputes": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Disputes",
        "description": "Every trade on hold, oldest first: the money waits in escrow until an admin resolves it.",
        "operationId": "disputes_admin_disputes_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/DisputeOut"
                  },
                  "type": "array",
                  "title": "Response Disputes Admin Disputes Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/transporters": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Transporters",
        "operationId": "transporters_admin_transporters_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/TransporterOut"
                  },
                  "type": "array",
                  "title": "Response Transporters Admin Transporters Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Add Transporter",
        "operationId": "add_transporter_admin_transporters_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TransporterIn"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TransporterOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/routes": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Route",
        "description": "The day's booked pickups in a suggested visiting order (Indian calendar day).",
        "operationId": "route_admin_routes_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "day",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date",
              "title": "Day"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RouteOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/anchors": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Anchors",
        "operationId": "anchors_admin_anchors_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/AnchorOut"
                  },
                  "type": "array",
                  "title": "Response Anchors Admin Anchors Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Seal",
        "operationId": "seal_admin_anchors_post",
        "responses": {
          "201": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnchorOut"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/anchors/{anchor_id}/check": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "Check",
        "operationId": "check_admin_anchors__anchor_id__check_get",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "anchor_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "title": "Anchor Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/AnchorCheckOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/admin/jobs": {
      "get": {
        "tags": [
          "operations"
        ],
        "summary": "List Jobs",
        "operationId": "list_jobs_admin_jobs_get",
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/JobOut"
                  },
                  "type": "array",
                  "title": "Response List Jobs Admin Jobs Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ]
      }
    },
    "/admin/jobs/{name}/run": {
      "post": {
        "tags": [
          "operations"
        ],
        "summary": "Run Now",
        "operationId": "run_now_admin_jobs__name__run_post",
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "parameters": [
          {
            "name": "name",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Name"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/JobRunOut"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "AnchorCheckOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "holds": {
            "type": "boolean",
            "title": "Holds"
          }
        },
        "type": "object",
        "required": [
          "id",
          "holds"
        ],
        "title": "AnchorCheckOut"
      },
      "AnchorOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "first_event_id": {
            "type": "integer",
            "title": "First Event Id"
          },
          "last_event_id": {
            "type": "integer",
            "title": "Last Event Id"
          },
          "event_count": {
            "type": "integer",
            "title": "Event Count"
          },
          "merkle_root": {
            "type": "string",
            "title": "Merkle Root"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "first_event_id",
          "last_event_id",
          "event_count",
          "merkle_root",
          "created_at"
        ],
        "title": "AnchorOut"
      },
...
      "DisputeInfoOut": {
        "properties": {
          "raised_by": {
            "type": "string",
            "enum": [
              "seller",
              "escrow"
            ],
            "title": "Raised By"
          },
          "reason": {
            "type": "string",
            "title": "Reason"
          },
          "raised_at": {
            "type": "string",
            "title": "Raised At"
          }
        },
        "type": "object",
        "required": [
          "raised_by",
          "reason",
          "raised_at"
        ],
        "title": "DisputeInfoOut"
      },
      "DisputeOut": {
        "properties": {
          "lot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lot Id"
          },
          "material_name": {
            "type": "string",
            "title": "Material Name"
          },
          "seller": {
            "$ref": "#/components/schemas/PartyOut"
          },
          "buyer": {
            "$ref": "#/components/schemas/PartyOut"
          },
          "dispute": {
            "$ref": "#/components/schemas/DisputeInfoOut"
          },
          "rate_paise_per_kg": {
            "type": "integer",
            "title": "Rate Paise Per Kg"
          },
          "declared_weight_grams": {
            "type": "integer",
            "title": "Declared Weight Grams"
          },
          "measured_weight_grams": {
            "type": "integer",
            "title": "Measured Weight Grams"
          },
          "escrow_held_paise": {
            "type": "integer",
            "title": "Escrow Held Paise"
          },
          "max_settle_weight_grams": {
            "type": "integer",
            "title": "Max Settle Weight Grams"
          }
        },
        "type": "object",
        "required": [
          "lot_id",
          "material_name",
          "seller",
          "buyer",
          "dispute",
          "rate_paise_per_kg",
          "declared_weight_grams",
          "measured_weight_grams",
          "escrow_held_paise",
          "max_settle_weight_grams"
        ],
        "title": "DisputeOut"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "JobOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "schedule": {
            "type": "string",
            "title": "Schedule"
          },
          "description": {
            "type": "string",
            "title": "Description"
          },
          "last_run": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/JobRunOut"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "type": "object",
        "required": [
          "name",
          "schedule",
          "description",
          "last_run"
        ],
        "title": "JobOut"
      },
      "JobRunOut": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "started_at": {
            "type": "string",
            "format": "date-time",
            "title": "Started At"
          },
          "finished_at": {
            "type": "string",
            "format": "date-time",
            "title": "Finished At"
          },
          "ok": {
            "type": "boolean",
            "title": "Ok"
          },
          "summary": {
            "type": "string",
            "title": "Summary"
          }
        },
        "type": "object",
        "required": [
          "name",
          "started_at",
          "finished_at",
          "ok",
          "summary"
        ],
        "title": "JobRunOut"
      },
...
      "LocationOut": {
        "properties": {
          "latitude": {
            "type": "number",
            "title": "Latitude"
          },
          "longitude": {
            "type": "number",
            "title": "Longitude"
          }
        },
        "type": "object",
        "required": [
          "latitude",
          "longitude"
        ],
        "title": "LocationOut"
      },
...
      "PartyOut": {
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "business_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Name"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "business_name"
        ],
        "title": "PartyOut"
      },
...
      "RouteOut": {
        "properties": {
          "day": {
            "type": "string",
            "format": "date",
            "title": "Day"
          },
          "depot": {
            "$ref": "#/components/schemas/LocationOut"
          },
          "stops": {
            "items": {
              "$ref": "#/components/schemas/RouteStopOut"
            },
            "type": "array",
            "title": "Stops"
          },
          "unplaced": {
            "items": {
              "$ref": "#/components/schemas/RouteStopOut"
            },
            "type": "array",
            "title": "Unplaced"
          },
          "return_metres": {
            "type": "integer",
            "title": "Return Metres"
          },
          "total_metres": {
            "type": "integer",
            "title": "Total Metres"
          }
        },
        "type": "object",
        "required": [
          "day",
          "depot",
          "stops",
          "unplaced",
          "return_metres",
          "total_metres"
        ],
        "title": "RouteOut"
      },
      "RouteStopOut": {
        "properties": {
          "lot_id": {
            "type": "string",
            "format": "uuid",
            "title": "Lot Id"
          },
          "seller": {
            "$ref": "#/components/schemas/PartyOut"
          },
          "material_name": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Material Name"
          },
          "pickup_at": {
            "type": "string",
            "format": "date-time",
            "title": "Pickup At"
          },
          "location": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/LocationOut"
              },
              {
                "type": "null"
              }
            ]
          },
          "leg_metres": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Leg Metres"
          }
        },
        "type": "object",
        "required": [
          "lot_id",
          "seller",
          "material_name",
          "pickup_at",
          "location",
          "leg_metres"
        ],
        "title": "RouteStopOut"
      },
...
      "TransporterIn": {
        "properties": {
          "name": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "pattern": "^\\+?[0-9]{10,15}$",
            "title": "Phone"
          },
          "vehicle": {
            "type": "string",
            "maxLength": 120,
            "minLength": 1,
            "title": "Vehicle"
          },
          "capacity_grams": {
            "type": "integer",
            "maximum": 100000000.0,
            "exclusiveMinimum": 0.0,
            "title": "Capacity Grams"
          }
        },
        "type": "object",
        "required": [
          "name",
          "phone",
          "vehicle",
          "capacity_grams"
        ],
        "title": "TransporterIn"
      },
      "TransporterOut": {
        "properties": {
          "id": {
            "type": "integer",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "phone": {
            "type": "string",
            "title": "Phone"
          },
          "vehicle": {
            "type": "string",
            "title": "Vehicle"
          },
          "capacity_grams": {
            "type": "integer",
            "title": "Capacity Grams"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "phone",
          "vehicle",
          "capacity_grams",
          "created_at"
        ],
        "title": "TransporterOut"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  }
}